# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Dijigoo Kurye Mobil API
  version: 1.0.0-draft.2
  description: |-
    Dijigoo Kurye saha uygulamasinin mobil API sozlesmesi.

    Iki urun, bir kimlik: web (kurye.dijigoo.com) onboarding tasir;
    mobil saha operasyonunu tasir. Ayni Next.js ailesinde yasir.

    ## Namespace

    - Mobil token API: `https://kurye.dijigoo.com/api/mobile` + bu belgedeki `/v1/*`
      (tam yol `/api/mobile/v1/...`).
    - Web cookie session (`dijigoo_courier_session`) dokunulmaz; mobil kullanmaz.
    - Web public API (`/api/public/v1/courier-*`) kimlik kaynagi olarak devralinir.
    - Iki OTP havuzu: kurye aktivasyon (`courierId`) ≠ alici teslim (`taskId+phone`).

    ## Sozlesme kurallari

    - Tum tarih alanlari RFC 3339, UTC ve acik offsetli.
    - Tum mutasyon istekleri `Idempotency-Key` basligi tasir. Ayni anahtarla
      gelen tekrar istegi ilk yanitin aynisini doner.
    - Gorev mutasyonlari `rowVersion` tasir; uyusmazlik 409 VERSION_MISMATCH.
    - Hata govdesi her zaman `application/problem+json` ve `ErrorResponse` semasidir.
    - Sayfalama imlec tabanlidir; `nextCursor` null ise son sayfadir.
    - Surumleme yol tabanlidir (`/v1`). Kirici degisiklik yeni yol acar.
    - Workflow adim tipleri kapali enum. v1 renderer: PHOTO_EVIDENCE ve FORM.select.
    - Vardiya fotografi yuz eslestirme yapmaz (KVKK m.6); `shiftFaceMatch=false`.
  contact:
    name: Dijigoo teknik ekip
servers:
  - url: https://kurye.dijigoo.com/api/mobile
...
    description: Yerel panel
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Aktivasyon sonrasi verilen kisa omurlu access token. 15 dakika gecerli, refresh ile
        doner. Token cihaz kimligine (installationId) baglidir.
    serviceAuth:
      type: apiKey
      in: header
      name: x-service-token
      description: Operasyon panelinin backend’i icin paylasilan servis anahtari. Kurye bearer tokeni
        degildir, bu API’de operator kimligi yoktur.
    solvelineInboundAuth:
      type: http
      scheme: bearer
      description: Solveline sesli asistanin bizim IVR sorgu/ticket API’sini cagirmasi icin ayri servis
        tokeni (SOLVELINE_INBOUND_TOKEN). Kurye JWT ve panel x-service-token degildir.
  schemas:
...
    TaskMutationResponse:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/TaskDetail"
        appliedAt:
          type: string
          format: date-time
          example: 2026-08-25T09:15:00.000Z
        replayed:
          type: boolean
          default: false
      required:
        - task
        - appliedAt
...
paths:
...
  /v1/tasks/{taskId}/transition:
    post:
      tags:
        - Task
      summary: Gorev durumunu ilerlet
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            example: 9f1c2f8a-7d1e-4f6b-9a3c-2b5d4e6f7a81
          required: true
          name: taskId
          in: path
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskTransitionRequest"
      responses:
        "200":
          description: Durum guncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskMutationResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/tasks/{taskId}/steps:
    post:
      tags:
        - Task
      summary: Workflow adimi gonder
      description: workflowVersion gorevin baslatildigi surumle eslesmelidir. Panel yeni surum yayinlamis
        olsa bile calisan gorev basladigi surumle biter.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            example: 9f1c2f8a-7d1e-4f6b-9a3c-2b5d4e6f7a81
          required: true
          name: taskId
          in: path
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StepSubmitRequest"
      responses:
        "200":
          description: Adim kaydedildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskMutationResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/tasks/{taskId}/finalize:
    post:
      tags:
        - Task
      summary: Gorevi sonuclandir (teslim edildi / edilemedi)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
            example: 9f1c2f8a-7d1e-4f6b-9a3c-2b5d4e6f7a81
          required: true
          name: taskId
          in: path
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskFinalizeRequest"
      responses:
        "200":
          description: Gorev sonuclandi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskMutationResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/custody/handover:
    post:
      tags:
        - Custody
      summary: Zimmet devret veya devral
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustodyHandoverRequest"
      responses:
        "200":
          description: Devir tamam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustodyHandoverResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/support/tickets:
    get:
      tags:
        - Support
      summary: Destek kayitlari
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type:
              - string
              - "null"
            description: Onceki yanittaki nextCursor
          required: false
          description: Onceki yanittaki nextCursor
          name: cursor
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          required: false
          name: limit
          in: query
        - schema:
            type:
              - string
              - "null"
            maxLength: 80
          required: false
          name: barcode
          in: query
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
      responses:
        "200":
          description: Kayit listesi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicketListResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Support
      summary: Destek kaydi ac
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
        - schema:
            type: string
            format: uuid
            description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
              istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          required: true
          description: Her mutasyon istegi icin istemcinin urettigi UUID. Ayni anahtarla gelen tekrar
            istekleri ilk yanitin aynisini doner. Offline kuyrukta zorunlu.
          name: idempotency-key
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportTicketCreateRequest"
      responses:
        "201":
          description: Kayit acildi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportTicket"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Surum catismasi veya kayit zaten sonuclandirilmis
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Is kurali ihlali veya eksik kanit
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/sync/batch:
    post:
      tags:
        - Sync
      summary: Offline kuyrugu bosalt
      description: Tekil bir olayin reddedilmesi partiyi durdurmaz; her olay icin ayri sonuc doner.
        Istemci sadece yeniden denenebilir hatalarda tekrar gonderir.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            pattern: ^dijigoo-courier\/\d+\.\d+\.\d+ \((android|ios); build \d+\)$
            example: dijigoo-courier/1.0.0 (android; build 42)
          required: true
          name: x-client-info
          in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncBatchRequest"
      responses:
        "200":
          description: Olay bazli sonuclar
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncBatchResponse"
        "400":
          description: Gecersiz istek
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Kimlik dogrulanamadi veya token suresi doldu
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Yetkisiz, cihaz bagli degil veya vardiya acik degil
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Kayit bulunamadi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Hiz siniri asildi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Beklenmeyen sunucu hatasi
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...